requirements.txt
attrs==25.4.0
blinker==1.9.0
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
dash==3.4.0
dash-bootstrap-components==2.0.4
exceptiongroup==1.3.1
fastf1==3.7.0
Flask==3.1.2
fonttools==4.60.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
msgpack==1.0.2
narwhals==2.15.0
nest-asyncio==1.6.0
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
plotly==6.5.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
RapidFuzz==3.13.0
requests==2.32.5
requests-cache==1.2.1
retrying==1.4.2
scipy==1.13.1
signalrcore==0.9.71
six==1.17.0
timple==0.1.8
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.5
zipp==3.23.0
